+++ packages/cra-template-seed/template/server/data/User/GET_USER_LOGGED.js
@@ -1,0 +1,5 @@
+module.exports = {
+  id: 'UG9rZW1vbjowMDE=',
+  name: 'John Doe',
+  permissions: ['canAccessTodoListPage', 'canAccessTodoEditPage', 'canAccessDropApplication'],
+};