+++ packages/main/rollup.config.js
@@ -1,3 +1,0 @@
-const rollupConfigFactory = require('../../scripts/rollup/configFactory');
-
-module.exports = rollupConfigFactory('main', ['@ui5/webcomponents-base']);