+++ packages/main/src/components/FlexBox/FlexBox.jss.ts
@@ -1,0 +1,60 @@
+export const styles = {
+  flexBox: {
+    display: 'flex'
+  },
+  fitContainer: { width: '100%', height: '100%' },
+  flexBoxDisplayInline: {
+    display: 'inline-flex'
+  },
+  justifyContentStart: {
+    justifyContent: 'flex-start'
+  },
+  justifyContentEnd: {
+    justifyContent: 'flex-end'
+  },
+  justifyContentCenter: {
+    justifyContent: 'center'
+  },
+  justifyContentSpaceAround: {
+    justifyContent: 'space-around'
+  },
+  justifyContentSpaceBetween: {
+    justifyContent: 'space-between'
+  },
+  flexBoxDirectionColumn: {
+    flexDirection: 'column'
+  },
+  flexBoxDirectionColumnReverse: {
+    flexDirection: 'column-reverse'
+  },
+  flexBoxDirectionRow: {
+    flexDirection: 'row'
+  },
+  flexBoxDirectionRowReverse: {
+    flexDirection: 'row-reverse'
+  },
+  alignItemsStart: {
+    alignItems: 'flex-start'
+  },
+  alignItemsEnd: {
+    alignItems: 'flex-end'
+  },
+  alignItemsCenter: {
+    alignItems: 'center'
+  },
+  alignItemsStretch: {
+    alignItems: 'stretch'
+  },
+  alignItemsBaseline: {
+    alignItems: 'baseline'
+  },
+  flexWrapNoWrap: {
+    flexWrap: 'nowrap'
+  },
+  flexWrapWrap: {
+    flexWrap: 'wrap'
+  },
+  flexWrapWrapReverse: {
+    flexWrap: 'wrap-reverse'
+  }
+};