+++ packages/main/src/webComponents/Tokenizer/Tokenizer.test.tsx
@@ -1,10 +1,10 @@
+import { render } from '@shared/tests';
 import { Tokenizer } from '@ui5/webcomponents-react/lib/Tokenizer';
-import { mount } from 'enzyme';
 import React from 'react';
 
 describe('Tokenizer', () => {
   test('Basic Test (generated)', () => {
-    const wrapper = mount(<Tokenizer />);
-    expect(wrapper.render()).toMatchSnapshot();
+    const { asFragment } = render(<Tokenizer />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });